Android - Add APKs.
I tested the "app-release.apk" on my phone and it doesn't seem
to work, even though it was tested and it worked on an emulator.
I checked the permissions and it seems that it has Internet access
permission (again, it worked on an emulator). I also checked that
the HTTP server and my phone are on the same network - I even
managed to `curl` an HTTP GET request from my phone to the
HTTP server. We'll probably demonstrate it on an emulator anyway,
and it seems like it's going to be a time waster trying to fix it,
so I'm probably going to leave it as is.